+++ packages/coast_audio/lib/coast_audio.dart
@@ -14,7 +14,6 @@ export 'src/data_source/abstract/audio_input_data_source.dart';
 export 'src/data_source/abstract/audio_output_data_source.dart';
 export 'src/data_source/audio_file_data_source.dart';
 export 'src/data_source/audio_memory_data_source.dart';
-export 'src/data_source/seek_origin.dart';
 export 'src/decoder/audio_decoder.dart';
 export 'src/decoder/wav/wav_audio_decoder.dart';
 export 'src/encoder/audio_encoder.dart';