+++ packages/coast_audio/lib/src/encoder/audio_encoder.dart
@@ -1,5 +1,8 @@
 import 'package:coast_audio/coast_audio.dart';
 
+/// The result of [AudioEncoder.encode].
+///
+/// [frames] is the number of frames encoded.
 class AudioEncodeResult {
   const AudioEncodeResult({
     required this.frames,
@@ -7,12 +10,21 @@ class AudioEncodeResult {
   final int frames;
 }
 
+/// An abstract class for audio encoders.
 abstract class AudioEncoder {
+  /// Input format of the encoder.
+  ///
+  /// Call the [encode] method with an [AudioBuffer] in this format.
   AudioFormat get inputFormat;
 
+  /// Starts encoding.
+  ///
+  /// Call this method before calling [encode].
   void start();
 
+  /// Encodes the audio from [buffer].
   AudioEncodeResult encode(AudioBuffer buffer);
 
+  /// Finishes encoding.
   void finalize();
 }