+++ packages/coast_audio/test/ring_buffer/ring_buffer_test.dart
@@ -32,29 +32,4 @@ void main() {
 
     expect(outputListView, List.generate(100, (i) => i));
   });
-
-  test('output == input overflow', () {
-    final pBuffer = malloc.allocate(10);
-    final ringBuffer = RingBuffer(capacity: 10, pBuffer: pBuffer.cast());
-
-    final pInput = malloc.allocate(100);
-    final pOutput = malloc.allocate(100);
-
-    final inputListView = pInput.cast<Uint8>().asTypedList(100);
-    final outputListView = pOutput.cast<Uint8>().asTypedList(100);
-
-    for (var i = 0; 100 > i; i++) {
-      inputListView[i] = i;
-    }
-
-    for (var i = 0; 10 > i; i++) {
-      final offset = i * 10;
-      ringBuffer.write(pInput.cast(), offset, 3); // length: 3, read: 0
-      ringBuffer.write(pInput.cast(), offset + 3, 6); // length: 9, read: 0
-      ringBuffer.write(pInput.cast(), offset + 9, 5); // length: 10, read: 0
-      ringBuffer.read(pOutput.cast(), offset, 10); // length: 0, read: 10
-    }
-
-    expect(outputListView, List.generate(100, (i) => i));
-  });
 }