+++ packages/coast_audio/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.0.4
+
+- Improve RingBuffer and FrameRingBuffer performance.
+- Improve FfiMemory performance.
+- Add documentation.
+
+### Breaking Changes
+
+- Set minimum Dart language version to 3.0.0.
+- Remove offset and count parameters from AudioInputDataSource.readBytes and AudioOutputDataSource.writeBytes.
+- Replace seek method with position property in AudioInputDataSource and AudioOutputDataSource.
+- Remove SeekOrigin.
+
 ## 0.0.3
 
 - Tweak AudioDecoder API.